Test for getting a retrieval for services
