+++ Calligraphy.Data/Repo/Service/ServiceRepoImpl.cs
@@ -1,4 +1,6 @@
-﻿using Calligraphy.Data.Models;
+﻿using Calligraphy.Data.Config;
+using Calligraphy.Data.Models;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -7,6 +9,20 @@ namespace Calligraphy.Data.Repo.Service
 {
     public class ServiceRepoImpl : IServiceRepo
     {
+        private readonly CalligraphyContext _context;
+        private readonly DbContextOptions<CalligraphyContext> options;
+
+
+        public ServiceRepoImpl(CalligraphyContext context)
+        {
+            _context = context;
+        }
+
+        public ServiceRepoImpl()
+        {
+            _context = new CalligraphyContext(options);
+        }
+
         public IEnumerable<ServiceEntity> GetAll()
         {
             throw new NotImplementedException();
